refactor(frontend): rename slot in Card for amount sub-text
