+++ src/frontend/src/lib/components/tokens/TokenCard.svelte
@@ -17,5 +17,5 @@
 
 	<slot name="balance" slot="amount" />
 
-	<slot name="exchange" slot="subAmount" />
+	<slot name="exchange" slot="amountDescription" />
 </Card>
