+++ src/frontend/src/lib/components/ui/CardAmountDescription.svelte
